neural-compression-pipeline/service_ocr/requirements.txt
fastapi
uvicorn
celery
redis
pillow
numpy
tensorflow==2.13.0
opencv-python-headless
lmdb
httpx>=0.27
python-multipart
pytesseract
Pillow
